Rewrite to computations
This is a major shift for the library.
Now,
instead of composing functions to implement algorithms,
algorithms are implemented as abstract representations of computations.
Those computations can then be run,
introspected,
or transformed in other ways.

Currently,
only running without compilation is supported.
However,
the framework is flexible
and can support other transformations
or ways of running,
such as running on a GPU.

The computation-framework is still a work in progress,
and breaking changes are expected.